Add '--persistent-srk' to make SRK persistent
This commit adds the '--persistent-srk' option to the
'pcr-oracle seal-secret' command, allowing the SRK to be stored at a
persistent handle in the TPM's non-volatile memory.

By making the SRK persistent, the bootloader can skip the time-consuming
key generation step on subsequent boots to reduce startup time.

The option will overwrite any existing object at the specified handle
and update the parent field in the output TPM 2.0 key file to reference
this new persistent location.

NOTE: This option doesn't support systemd-boot yet.

+++ src/tpm2key.c
@@ -53,7 +53,8 @@ tpm2key_basekey(TSSPRIVKEY **tpm2key, const TPM2_HANDLE parent,
 	key->type = OBJ_txt2obj(OID_sealedData, 1);
 	key->emptyAuth = 1;
 	key->parent = ASN1_INTEGER_new();
-	ASN1_INTEGER_set(key->parent, parent);
+	/* Convert 'parent' to 'int32_t' to make sure 'parent' only occupies at most 4 bytes */
+	ASN1_INTEGER_set(key->parent, (int32_t)parent);
 
 	key->pubkey = ASN1_OCTET_STRING_new();
 	ASN1_STRING_set(key->pubkey, bp_pub->data, buffer_available(bp_pub));
